examples/aiomysql-mysql/requirements.txt
dbop-core[aiomysql]
python-dotenv
cryptography>=42